feat(runtime-c-api) Finish to extract last pieces of code.
